Added data to be shown
The country, state, active, active cases and the last time it was
updated are now shown on the web in addition to the previous data.
I also further deleted some of the data files, when we actually start
plotting the data we will introduce some granularity that makes it so we
can handle such large amounts of data, but for now I will use a smaller
dataset.